email_agent/requirements.txt
# Core LangChain and LangGraph
langchain>=1.0.0,<2.0.0
langchain-core>=1.0.0,<2.0.0
langchain-openai>=1.0.0,<2.0.0
langgraph>=1.0.0,<2.0.0
langsmith>=0.4.37
langgraph-cli[inmem]>=0.4.0
langchain-mcp-adapters>=0.1.12

# Azure Services
azure-identity>=1.15.0
langchain-azure-ai>=1.0.0
langchain-azure-storage>=1.0.0
foundry-local-sdk
langchain-azure-postgresql>=1.0.0 

# Database
psycopg[binary]>=3.1.0
pgvector>=0.2.0

# MCP (Model Context Protocol)
# Note: Install @softeria/ms-365-mcp-server via npm
# npm install -g @softeria/ms-365-mcp-server

# Web API
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.0.0

# Utilities
python-dotenv
html2text
rich
